pyproject.toml
[project]
name = "config-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "twisted",
    "klein",
    "psycopg2-binary",
    "pyyaml",
    "jinja2",
    "dependency-injector",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-twisted",
    "mypy",
    "ruff",
    "types-PyYAML",
]

[project.scripts]
config-service = "config_service.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/config_service"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
